+++ scheds/rust/scx_lavd/src/bpf/main.bpf.c
@@ -182,6 +182,7 @@
  */
 #include <scx/common.bpf.h>
 #include "intf.h"
+#include "lavd.bpf.h"
 #include <errno.h>
 #include <stdbool.h>
 #include <string.h>
@@ -192,34 +193,14 @@
 char _license[] SEC("license") = "GPL";
 
 /*
- * Sched related globals
+ * Include sub-modules
  */
-struct sys_stat	__sys_stats[2];
-volatile int	__sys_stat_idx;
 
-private(LAVD) struct bpf_cpumask __kptr *turbo_cpumask; /* CPU mask for turbo CPUs */
-private(LAVD) struct bpf_cpumask __kptr *big_cpumask; /* CPU mask for big CPUs */
-private(LAVD) struct bpf_cpumask __kptr *little_cpumask; /* CPU mask for little CPUs */
-private(LAVD) struct bpf_cpumask __kptr *active_cpumask; /* CPU mask for active CPUs */
-private(LAVD) struct bpf_cpumask __kptr *ovrflw_cpumask; /* CPU mask for overflow CPUs */
-private(LAVD) struct bpf_cpumask cpdom_cpumask[LAVD_CPDOM_MAX_NR]; /* CPU mask for each compute domain */
-
-static u64		LAVD_AP_LOW_UTIL;
-static bool		have_turbo_core;
-static bool		have_little_core;
-
-/*
- * CPU topology
- */
-const volatile u64	nr_cpu_ids;	/* maximum CPU IDs */
-static volatile u64	nr_cpus_onln;	/* current number of online CPUs */
-static volatile u64	nr_cpus_big;
-
-const volatile u16	cpu_order_performance[LAVD_CPU_ID_MAX]; /* CPU preference order for performance and balanced mode */
-const volatile u16	cpu_order_powersave[LAVD_CPU_ID_MAX]; /* CPU preference order for powersave mode */
-const volatile u16	__cpu_capacity_hint[LAVD_CPU_ID_MAX]; /* CPU capacity based on 1000 */
-struct cpdom_ctx	cpdom_ctxs[LAVD_CPDOM_MAX_NR]; /* contexts for compute domains */
-
+#include "util.bpf.c"
+#include "introspec.bpf.c"
+#include "power.bpf.c"
+#include "sys_stat.bpf.c"
+#include "preempt.bpf.c"
 
 /*
  * Logical current clock
@@ -230,1029 +211,6 @@ static u64		cur_logical_clk;
  * Current service time
  */
 static u64		cur_svc_time;
-
-/*
- * Big core's compute ratio among currently active cores
- */
-static u32		cur_big_core_ratio;
-
-/*
- * Big core's compute ratio when all cores are active
- */
-static u32		default_big_core_ratio;
-
-/*
- * Options
- */
-volatile bool		no_core_compaction;
-volatile bool		no_freq_scaling;
-volatile bool		no_prefer_turbo_core;
-volatile bool		is_powersave_mode;
-volatile bool		reinit_cpumask_for_performance;
-const volatile bool	is_autopilot_on;
-const volatile u32 	is_smt_active;
-const volatile u8	verbose;
-
-/*
- * Statistics
- */
-volatile int		power_mode;
-volatile u64		last_power_mode_clk;
-volatile u64		performance_mode_ns;
-volatile u64		balanced_mode_ns;
-volatile u64		powersave_mode_ns;
-
-/*
- * Exit infomation
- */
-UEI_DEFINE(uei);
-
-#define debugln(fmt, ...)						\
-({									\
-	if (verbose > 0)						\
-		bpf_printk("[%s:%d] " fmt, __func__, __LINE__,		\
-					##__VA_ARGS__);			\
-})
-
-#define traceln(fmt, ...)						\
-({									\
-	if (verbose > 1)						\
-		bpf_printk("[%s:%d] " fmt, __func__, __LINE__,		\
-					##__VA_ARGS__);			\
-})
-
-#ifndef min
-#define min(X, Y) (((X) < (Y)) ? (X) : (Y))
-#endif
-
-#ifndef max
-#define max(X, Y) (((X) < (Y)) ? (Y) : (X))
-#endif
-
-/*
- * Timer for updating system-wide CPU utilization periorically
- */
-struct update_timer {
-	struct bpf_timer timer;
-};
-
-struct {
-	__uint(type, BPF_MAP_TYPE_ARRAY);
-	__uint(max_entries, 1);
-	__type(key, u32);
-	__type(value, struct update_timer);
-} update_timer SEC(".maps");
-
-/*
- * per-CPU globals
- */
-struct {
-	__uint(type, BPF_MAP_TYPE_PERCPU_ARRAY);
-	__type(key, u32);
-	__type(value, struct cpu_ctx);
-	__uint(max_entries, 1);
-} cpu_ctx_stor SEC(".maps");
-
-/*
- * Per-task scheduling context
- */
-struct {
-	__uint(type, BPF_MAP_TYPE_TASK_STORAGE);
-	__uint(map_flags, BPF_F_NO_PREALLOC);
-	__type(key, int);
-	__type(value, struct task_ctx);
-} task_ctx_stor SEC(".maps");
-
-/*
- * Preemption related ones
- */
-struct preemption_info {
-	u64		stopping_tm_est_ns;
-	u64		last_kick_clk;
-	u64		lat_cri;
-	struct cpu_ctx	*cpuc;
-};
-
-/*
- * Introspection commands
- */
-struct introspec intrspc;
-
-struct {
-	__uint(type, BPF_MAP_TYPE_RINGBUF);
-	__uint(max_entries, 16 * 1024 /* 16 KB */);
-} introspec_msg SEC(".maps");
-
-static u16 get_nice_prio(struct task_struct *p);
-static int reinit_active_cpumask_for_performance(void);
-static void update_power_mode_time(void);
-
-static u64 sigmoid_u64(u64 v, u64 max)
-{
-	/*
-	 * An integer approximation of the sigmoid function. It is convenient
-	 * to use the sigmoid function since it has a known upper and lower
-	 * bound, [0, max].
-	 *
-	 *      |
-	 *	|      +------ <= max
-	 *	|    /
-	 *	|  /
-	 *	|/
-	 *	+------------->
-	 */
-	return (v > max) ? max : v;
-}
-
-static u64 rsigmoid_u64(u64 v, u64 max)
-{
-	/*
-	 * A horizontally flipped version of sigmoid function. Again, it is
-	 * convenient since the upper and lower bound of the function is known,
-	 * [0, max].
-	 *
-	 *
-	 *      |
-	 *	|\ <= max
-	 *	| \
-	 *	|  \
-	 *	|   \
-	 *	+----+-------->
-	 */
-	return (v >= max) ? 0 : max - v;
-}
-
-static struct task_ctx *try_get_task_ctx(struct task_struct *p)
-{
-	return bpf_task_storage_get(&task_ctx_stor, p, 0, 0);
-}
-
-static struct task_ctx *get_task_ctx(struct task_struct *p)
-{
-	struct task_ctx *taskc;
-
-	taskc = try_get_task_ctx(p);
-	if (!taskc)
-		scx_bpf_error("task_ctx lookup failed for %s[%d]",
-			      p->comm, p->pid);
-	return taskc;
-}
-
-static struct cpu_ctx *get_cpu_ctx(void)
-{
-	const u32 idx = 0;
-	struct cpu_ctx *cpuc;
-
-	cpuc = bpf_map_lookup_elem(&cpu_ctx_stor, &idx);
-	if (!cpuc)
-		scx_bpf_error("cpu_ctx lookup failed for current cpu");
-
-	return cpuc;
-}
-
-static struct cpu_ctx *get_cpu_ctx_id(s32 cpu_id)
-{
-	const u32 idx = 0;
-	struct cpu_ctx *cpuc;
-
-	cpuc = bpf_map_lookup_percpu_elem(&cpu_ctx_stor, &idx, cpu_id);
-	if (!cpuc)
-		scx_bpf_error("cpu_ctx lookup failed for %d", cpu_id);
-
-	return cpuc;
-}
-
-static struct sys_stat *get_sys_stat_cur(void)
-{
-	if (READ_ONCE(__sys_stat_idx) == 0)
-		return &__sys_stats[0];
-	return &__sys_stats[1];
-}
-
-static struct sys_stat *get_sys_stat_next(void)
-{
-	if (READ_ONCE(__sys_stat_idx) == 0)
-		return &__sys_stats[1];
-	return &__sys_stats[0];
-}
-
-static void flip_sys_stat(void)
-{
-	WRITE_ONCE(__sys_stat_idx, __sys_stat_idx ^ 0x1);
-}
-
-static bool is_lat_cri(struct task_ctx *taskc, struct sys_stat *stat_cur)
-{
-	return taskc->lat_cri >= stat_cur->avg_lat_cri;
-}
-
-static bool is_perf_cri(struct task_ctx *taskc, struct sys_stat *stat_cur)
-{
-	if (READ_ONCE(taskc->on_big) && READ_ONCE(taskc->on_little))
-		return taskc->perf_cri >= stat_cur->thr_perf_cri;
-	return READ_ONCE(taskc->on_big);
-}
-
-static bool is_greedy(struct task_ctx *taskc)
-{
-	return taskc->greedy_ratio > 1000;
-}
-
-static bool is_eligible(struct task_ctx *taskc)
-{
-	return !is_greedy(taskc);
-}
-
-
-static __always_inline
-int submit_task_ctx(struct task_struct *p, struct task_ctx *taskc, u32 cpu_id)
-{
-	struct sys_stat *stat_cur = get_sys_stat_cur();
-	struct cpu_ctx *cpuc;
-	struct msg_task_ctx *m;
-
-	cpuc = get_cpu_ctx_id(cpu_id);
-	if (!cpuc)
-		return -EINVAL;
-
-	m = bpf_ringbuf_reserve(&introspec_msg, sizeof(*m), 0);
-	if (!m)
-		return -ENOMEM;
-
-	m->hdr.kind = LAVD_MSG_TASKC;
-	m->taskc_x.pid = p->pid;
-	memcpy(m->taskc_x.comm, p->comm, TASK_COMM_LEN);
-	m->taskc_x.static_prio = get_nice_prio(p);
-	m->taskc_x.cpu_util = cpuc->util / 10;
-	m->taskc_x.cpu_id = cpu_id;
-	m->taskc_x.avg_lat_cri = stat_cur->avg_lat_cri;
-	m->taskc_x.thr_perf_cri = stat_cur->thr_perf_cri;
-	m->taskc_x.nr_active = stat_cur->nr_active;
-	m->taskc_x.cpuperf_cur = cpuc->cpuperf_cur;
-
-	m->taskc_x.stat[0] = is_lat_cri(taskc, stat_cur) ? 'L' : 'R';
-	m->taskc_x.stat[1] = is_perf_cri(taskc, stat_cur) ? 'H' : 'I';
-	m->taskc_x.stat[2] = cpuc->big_core ? 'B' : 'T';
-	m->taskc_x.stat[3] = is_greedy(taskc) ? 'G' : 'E';
-	m->taskc_x.stat[4] = taskc->victim_cpu >= 0 ? 'P' : 'N';
-	m->taskc_x.stat[5] = '\0';
-
-	memcpy(&m->taskc, taskc, sizeof(m->taskc));
-
-	bpf_ringbuf_submit(m, 0);
-
-	return 0;
-}
-
-static void proc_introspec_sched_n(struct task_struct *p,
-				   struct task_ctx *taskc, u32 cpu_id)
-{
-	u64 cur_nr, prev_nr;
-	int i;
-
-	/* introspec_arg is the number of schedules remaining */
-	cur_nr = intrspc.arg;
-
-	/*
-	 * Note that the bounded retry (@LAVD_MAX_RETRY) does *not *guarantee*
-	 * to decrement introspec_arg. However, it is unlikely to happen. Even
-	 * if it happens, it is nothing but a matter of delaying a message
-	 * delivery. That's because other threads will try and succeed the CAS
-	 * operation eventually. So this is good enough. ;-)
-	 */
-	for (i = 0; cur_nr > 0 && i < LAVD_MAX_RETRY; i++) {
-		prev_nr = __sync_val_compare_and_swap(
-				&intrspc.arg, cur_nr, cur_nr - 1);
-		/* CAS success: submit a message and done */
-		if (prev_nr == cur_nr) {
-			submit_task_ctx(p, taskc, cpu_id);
-			break;
-		}
-		/* CAS failure: retry */
-		cur_nr = prev_nr;
-	}
-}
-
-static void proc_introspec_pid(struct task_struct *p, struct task_ctx *taskc,
-			       u32 cpu_id)
-{
-	if (p->pid == intrspc.arg)
-		submit_task_ctx(p, taskc, cpu_id);
-}
-
-static bool have_scheduled(struct task_ctx *taskc)
-{
-	/*
-	 * If task's time slice hasn't been updated, that means the task has
-	 * been scheduled by this scheduler.
-	 */
-	return taskc->slice_ns != 0;
-}
-
-static void try_proc_introspec_cmd(struct task_struct *p,
-				   struct task_ctx *taskc, u32 cpu_id)
-{
-	if (LAVD_CPU_ID_HERE == cpu_id)
-		cpu_id = bpf_get_smp_processor_id();
-
-	switch(intrspc.cmd) {
-	case LAVD_CMD_SCHED_N:
-		proc_introspec_sched_n(p, taskc, cpu_id);
-		break;
-	case LAVD_CMD_NOP:
-		/* do nothing */
-		break;
-	default:
-		scx_bpf_error("Unknown introspec command: %d", intrspc.cmd);
-		break;
-	}
-}
-
-static u32 calc_avg32(u32 old_val, u32 new_val)
-{
-	/*
-	 * Calculate the exponential weighted moving average (EWMA).
-	 *  - EWMA = (0.75 * old) + (0.25 * new)
-	 */
-	return (old_val - (old_val >> 2)) + (new_val >> 2);
-}
-
-static u64 calc_avg(u64 old_val, u64 new_val)
-{
-	/*
-	 * Calculate the exponential weighted moving average (EWMA).
-	 *  - EWMA = (0.75 * old) + (0.25 * new)
-	 */
-	return (old_val - (old_val >> 2)) + (new_val >> 2);
-}
-
-static u64 calc_avg_freq(u64 old_freq, u64 interval)
-{
-	u64 new_freq, ewma_freq;
-
-	/*
-	 * Calculate the exponential weighted moving average (EWMA) of a
-	 * frequency with a new interval measured.
-	 */
-	new_freq = LAVD_TIME_ONE_SEC / interval;
-	ewma_freq = calc_avg(old_freq, new_freq);
-	return ewma_freq;
-}
-
-struct sys_stat_ctx {
-	struct sys_stat *stat_cur;
-	struct sys_stat	*stat_next;
-	u64		now;
-	u64		duration;
-	u64		duration_total;
-	u64		idle_total;
-	u64		compute_total;
-	u64		load_actual;
-	u64		tot_svc_time;
-	u64		nr_queued_task;
-	u64		load_run_time_ns;
-	s32		max_lat_cri;
-	s32		avg_lat_cri;
-	u64		sum_lat_cri;
-	u32		nr_sched;
-	u32		nr_migration;
-	u32		nr_preemption;
-	u32		nr_greedy;
-	u32		nr_perf_cri;
-	u32		nr_lat_cri;
-	u32		nr_big;
-	u32		nr_pc_on_big;
-	u32		nr_lc_on_big;
-	u64		min_perf_cri;
-	u64		avg_perf_cri;
-	u64		max_perf_cri;
-	u64		sum_perf_cri;
-	u32		thr_perf_cri;
-	u64		new_util;
-	u32		nr_violation;
-};
-
-static void init_sys_stat_ctx(struct sys_stat_ctx *c)
-{
-	memset(c, 0, sizeof(*c));
-
-	c->stat_cur = get_sys_stat_cur();
-	c->stat_next = get_sys_stat_next();
-	c->min_perf_cri = 1000;
-	c->now = bpf_ktime_get_ns();
-	c->duration = c->now - c->stat_cur->last_update_clk;
-	c->stat_next->last_update_clk = c->now;
-}
-
-static void collect_sys_stat(struct sys_stat_ctx *c)
-{
-	u64 dsq_id;
-	int cpu, nr;
-
-	bpf_for(cpu, 0, nr_cpu_ids) {
-		struct cpu_ctx *cpuc = get_cpu_ctx_id(cpu);
-		if (!cpuc) {
-			c->compute_total = 0;
-			break;
-		}
-
-		/*
-		 * Accumulate cpus' loads.
-		 */
-		c->load_actual += cpuc->load_actual;
-		c->load_run_time_ns += cpuc->load_run_time_ns;
-		c->tot_svc_time += cpuc->tot_svc_time;
-		cpuc->tot_svc_time = 0;
-
-		/*
-		 * Accumulate statistics.
-		 */
-		if (cpuc->big_core) {
-			c->nr_big += cpuc->nr_sched;
-			c->nr_pc_on_big += cpuc->nr_perf_cri;
-			c->nr_lc_on_big += cpuc->nr_lat_cri;
-		}
-		c->nr_perf_cri += cpuc->nr_perf_cri;
-		cpuc->nr_perf_cri = 0;
-
-		c->nr_lat_cri += cpuc->nr_lat_cri;
-		cpuc->nr_lat_cri = 0;
-
-		c->nr_migration += cpuc->nr_migration;
-		cpuc->nr_migration = 0;
-
-		c->nr_preemption += cpuc->nr_preemption;
-		cpuc->nr_preemption = 0;
-
-		c->nr_greedy += cpuc->nr_greedy;
-		cpuc->nr_greedy = 0;
-
-		/*
-		 * Accumulate task's latency criticlity information.
-		 *
-		 * While updating cpu->* is racy, the resulting impact on
-		 * accuracy should be small and very rare and thus should be
-		 * fine.
-		 */
-		c->sum_lat_cri += cpuc->sum_lat_cri;
-		cpuc->sum_lat_cri = 0;
-
-		c->nr_sched += cpuc->nr_sched;
-		cpuc->nr_sched = 0;
-
-		if (cpuc->max_lat_cri > c->max_lat_cri)
-			c->max_lat_cri = cpuc->max_lat_cri;
-		cpuc->max_lat_cri = 0;
-
-		/*
-		 * Accumulate task's performance criticlity information.
-		 */
-		if (cpuc->min_perf_cri < c->min_perf_cri)
-			c->min_perf_cri = cpuc->min_perf_cri;
-		cpuc->min_perf_cri = 1000;
-
-		if (cpuc->max_perf_cri > c->max_perf_cri)
-			c->max_perf_cri = cpuc->max_perf_cri;
-		cpuc->max_perf_cri = 0;
-
-		c->sum_perf_cri += cpuc->sum_perf_cri;
-		cpuc->sum_perf_cri = 0;
-
-		/*
-		 * If the CPU is in an idle state (i.e., idle_start_clk is
-		 * non-zero), accumulate the current idle peirod so far.
-		 */
-		for (int i = 0; i < LAVD_MAX_RETRY; i++) {
-			u64 old_clk = cpuc->idle_start_clk;
-			if (old_clk == 0)
-				break;
-
-			bool ret = __sync_bool_compare_and_swap(
-					&cpuc->idle_start_clk, old_clk, c->now);
-			if (ret) {
-				cpuc->idle_total += c->now - old_clk;
-				break;
-			}
-		}
-
-		/*
-		 * Calculcate per-CPU utilization
-		 */
-		u64 compute = 0;
-		if (c->duration > cpuc->idle_total)
-			compute = c->duration - cpuc->idle_total;
-
-		c->new_util = (compute * LAVD_CPU_UTIL_MAX) / c->duration;
-		cpuc->util = calc_avg(cpuc->util, c->new_util);
-
-		if (cpuc->turbo_core) {
-			if (cpuc->util > LAVD_CC_PER_TURBO_CORE_MAX_CTUIL)
-				c->nr_violation += 1000;
-		}
-		else {
-			if (cpuc->util > LAVD_CC_PER_CORE_MAX_CTUIL)
-				c->nr_violation += 1000;
-		}
-
-		/*
-		 * Accmulate system-wide idle time
-		 */
-		c->idle_total += cpuc->idle_total;
-		cpuc->idle_total = 0;
-	}
- 
-	bpf_for(dsq_id, 0, LAVD_CPDOM_MAX_NR) {
-		nr = scx_bpf_dsq_nr_queued(dsq_id);
-		if (nr > 0)
-			c->nr_queued_task += nr;
-	}
-}
-
-static void calc_sys_stat(struct sys_stat_ctx *c)
-{
-	c->duration_total = c->duration * nr_cpus_onln;
-	if (c->duration_total > c->idle_total)
-		c->compute_total = c->duration_total - c->idle_total;
-	else
-		c->compute_total = 0;
-	c->new_util = (c->compute_total * LAVD_CPU_UTIL_MAX)/c->duration_total;
-
-	if (c->nr_sched == 0) {
-		/*
-		 * When a system is completely idle, it is indeed possible
-		 * nothing scheduled for an interval.
-		 */
-		c->max_lat_cri = c->stat_cur->max_lat_cri;
-		c->avg_lat_cri = c->stat_cur->avg_lat_cri;
-
-		c->min_perf_cri = c->stat_cur->min_perf_cri;
-		c->max_perf_cri = c->stat_cur->max_perf_cri;
-		c->avg_perf_cri = c->stat_cur->avg_perf_cri;
-	}
-	else {
-		c->avg_lat_cri = c->sum_lat_cri / c->nr_sched;
-		c->avg_perf_cri = c->sum_perf_cri / c->nr_sched;
-	}
-}
-
-static void update_sys_stat_next(struct sys_stat_ctx *c)
-{
-	static int cnt = 0;
-	u64 avg_svc_time = 0;
-
-	/*
-	 * Update the CPU utilization to the next version.
-	 */
-	struct sys_stat *stat_cur = c->stat_cur;
-	struct sys_stat *stat_next = c->stat_next;
-
-	stat_next->load_actual =
-		calc_avg(stat_cur->load_actual, c->load_actual);
-	stat_next->util =
-		calc_avg(stat_cur->util, c->new_util);
-
-	stat_next->max_lat_cri =
-		calc_avg32(stat_cur->max_lat_cri, c->max_lat_cri);
-	stat_next->avg_lat_cri =
-		calc_avg32(stat_cur->avg_lat_cri, c->avg_lat_cri);
-	stat_next->thr_lat_cri = stat_next->max_lat_cri -
-		((stat_next->max_lat_cri - stat_next->avg_lat_cri) >> 1);
-
-	stat_next->min_perf_cri =
-		calc_avg32(stat_cur->min_perf_cri, c->min_perf_cri);
-	stat_next->avg_perf_cri =
-		calc_avg32(stat_cur->avg_perf_cri, c->avg_perf_cri);
-	stat_next->max_perf_cri =
-		calc_avg32(stat_cur->max_perf_cri, c->max_perf_cri);
-	stat_next->thr_perf_cri =
-		c->stat_cur->thr_perf_cri; /* will be updated later */
-
-	stat_next->nr_violation =
-		calc_avg32(stat_cur->nr_violation, c->nr_violation);
-
-	if (c->nr_sched > 0)
-		avg_svc_time = c->tot_svc_time / c->nr_sched;
-	stat_next->avg_svc_time =
-		calc_avg(stat_cur->avg_svc_time, avg_svc_time);
-
-	stat_next->nr_queued_task =
-		calc_avg(stat_cur->nr_queued_task, c->nr_queued_task);
-
-
-	/*
-	 * Half the statistics every minitue so the statistics hold the
-	 * information on a few minutes.
-	 */
-	if (cnt++ == LAVD_SYS_STAT_DECAY_TIMES) {
-		cnt = 0;
-		stat_next->nr_sched >>= 1;
-		stat_next->nr_migration >>= 1;
-		stat_next->nr_preemption >>= 1;
-		stat_next->nr_greedy >>= 1;
-		stat_next->nr_perf_cri >>= 1;
-		stat_next->nr_lat_cri >>= 1;
-		stat_next->nr_big >>= 1;
-		stat_next->nr_pc_on_big >>= 1;
-		stat_next->nr_lc_on_big >>= 1;
-
-		__sync_fetch_and_sub(&performance_mode_ns, performance_mode_ns/2);
-		__sync_fetch_and_sub(&balanced_mode_ns, balanced_mode_ns/2);
-		__sync_fetch_and_sub(&powersave_mode_ns, powersave_mode_ns/2);
-	}
-
-	stat_next->nr_sched += c->nr_sched;
-	stat_next->nr_migration += c->nr_migration;
-	stat_next->nr_preemption += c->nr_preemption;
-	stat_next->nr_greedy += c->nr_greedy;
-	stat_next->nr_perf_cri += c->nr_perf_cri;
-	stat_next->nr_lat_cri += c->nr_lat_cri;
-	stat_next->nr_big += c->nr_big;
-	stat_next->nr_pc_on_big += c->nr_pc_on_big;
-	stat_next->nr_lc_on_big += c->nr_lc_on_big;
-
-	update_power_mode_time();
-}
-
-static void do_update_sys_stat(void)
-{
-	struct sys_stat_ctx c;
-
-	/*
-	 * Collect and prepare the next version of stat.
-	 */
-	init_sys_stat_ctx(&c);
-	collect_sys_stat(&c);
-	calc_sys_stat(&c);
-	update_sys_stat_next(&c);
-
-	/*
-	 * Make the next version atomically visible.
-	 */
-	flip_sys_stat();
-}
-
-static u64 calc_nr_active_cpus(struct sys_stat *stat_cur)
-{
-	u64 nr_active;
-
-	/*
-	 * nr_active = ceil(nr_cpus_onln * cpu_util * per_core_max_util)
-	 */
-	nr_active  = (nr_cpus_onln * stat_cur->util * 1000) + 500;
-	nr_active /= (LAVD_CC_PER_CORE_MAX_CTUIL * 1000);
-
-	/*
-	 * If a few CPUs are particularly busy, boost the active CPUs more.
-	 */
-	nr_active += min(LAVD_CC_NR_OVRFLW, (stat_cur->nr_violation) / 1000);
-	nr_active = max(min(nr_active, nr_cpus_onln),
-			LAVD_CC_NR_ACTIVE_MIN);
-
-	return nr_active;
-}
-
-static bool clear_cpu_periodically(u32 cpu, struct bpf_cpumask *cpumask)
-{
-	u32 clear;
-
-	/*
-	 * If the CPU is on, we clear the bit once every four times
-	 * (LAVD_CC_CPU_PIN_INTERVAL_DIV). Hence, the bit will be
-	 * probabilistically cleared once every 100 msec (4 * 25 msec).
-	 */
-	clear = !(bpf_get_prandom_u32() % LAVD_CC_CPU_PIN_INTERVAL_DIV);
-	if (clear)
-		bpf_cpumask_clear_cpu(cpu, cpumask);
-
-	return clear;
-}
-
-static void do_core_compaction(void)
-{
-	struct sys_stat *stat_cur = get_sys_stat_cur();
-	struct cpu_ctx *cpuc;
-	struct bpf_cpumask *active, *ovrflw;
-	int nr_cpus, nr_active, nr_active_old, cpu, i;
-	u32 sum_capacity = 0, big_capacity = 0;
-	bool clear;
-	const volatile u16 *cpu_order;
-
-	bpf_rcu_read_lock();
-
-	/*
-	 * Prepare cpumasks.
-	 */
-	active = active_cpumask;
-	ovrflw = ovrflw_cpumask;
-	if (!active || !ovrflw) {
-		scx_bpf_error("Failed to prepare cpumasks.");
-		goto unlock_out;
-	}
-
-	/*
-	 * Decide a cpuorder to use according to its power mode.
-	 */
-	if (is_powersave_mode)
-		cpu_order = cpu_order_powersave;
-	else
-		cpu_order = cpu_order_performance;
-
-	/*
-	 * Assign active and overflow cores
-	 */
-	nr_active_old = stat_cur->nr_active;
-	nr_active = calc_nr_active_cpus(stat_cur);
-	nr_cpus = nr_active + LAVD_CC_NR_OVRFLW;
-	bpf_for(i, 0, nr_cpu_ids) {
-		if (i >= LAVD_CPU_ID_MAX)
-			break;
-
-		/*
-		 * Skip offline cpu
-		 */
-		cpu = cpu_order[i];
-		cpuc = get_cpu_ctx_id(cpu);
-		if (!cpuc || !cpuc->is_online) {
-			bpf_cpumask_clear_cpu(cpu, active);
-			bpf_cpumask_clear_cpu(cpu, ovrflw);
-			continue;
-		}
-
-		/*
-		 * Assign an online cpu to active and overflow cpumasks
-		 */
-		if (i < nr_cpus) {
-			if (i < nr_active) {
-				bpf_cpumask_set_cpu(cpu, active);
-				bpf_cpumask_clear_cpu(cpu, ovrflw);
-			}
-			else {
-				bpf_cpumask_set_cpu(cpu, ovrflw);
-				bpf_cpumask_clear_cpu(cpu, active);
-			}
-			scx_bpf_kick_cpu(cpu, SCX_KICK_IDLE);
-
-			/*
-			 * Calculate big capacity ratio among active cores.
-			 */
-			sum_capacity += cpuc->capacity;
-			if (cpuc->big_core)
-				big_capacity += cpuc->capacity;
-		}
-		else {
-			if (i < nr_active_old) {
-				bpf_cpumask_clear_cpu(cpu, active);
-				bpf_cpumask_clear_cpu(cpu, ovrflw);
-			}
-			else {
-				/*
-				 * This is the case when a CPU belongs to the
-				 * overflow set even though that CPU was not an
-				 * overflow set initially. This can happen only
-				 * when a pinned userspace task ran on this
-				 * CPU. In this case, we keep the CPU in an
-				 * overflow set since the CPU will be used
-				 * anyway for the task. This will promote equal
-				 * use of all used CPUs, lowering the energy
-				 * consumption by avoiding a few CPUs being
-				 * turbo-boosted. Hence, we do not clear the
-				 * overflow cpumask here for a while,
-				 * approximately for LAVD_CC_CPU_PIN_INTERVAL.
-				 */
-				bpf_cpumask_clear_cpu(cpu, active);
-				clear = clear_cpu_periodically(cpu, ovrflw);
-				if (!clear)
-					scx_bpf_kick_cpu(cpu, SCX_KICK_IDLE);
-			}
-		}
-	}
-
-	cur_big_core_ratio = (1000 * big_capacity) / sum_capacity;
-	stat_cur->nr_active = nr_active;
-
-unlock_out:
-	bpf_rcu_read_unlock();
-}
-
-static void update_power_mode_time(void)
-{
-	u64 now = bpf_ktime_get_ns();
-	u64 delta;
-
-	if (last_power_mode_clk == 0)
-		last_power_mode_clk = now;
-
-	delta = now - last_power_mode_clk;
-	last_power_mode_clk = now;
-
-	switch (power_mode) {
-	case LAVD_PM_PERFORMANCE:
-		__sync_fetch_and_add(&performance_mode_ns, delta);
-		break;
-	case LAVD_PM_BALANCED:
-		__sync_fetch_and_add(&balanced_mode_ns, delta);
-		break;
-	case LAVD_PM_POWERSAVE:
-		__sync_fetch_and_add(&powersave_mode_ns, delta);
-		break;
-	}
-}
-
-
-static int do_set_power_profile(s32 pm, int util)
-{
-	/*
-	 * Skip setting the mode if already in the same mode.
-	 */
-	if (power_mode == pm)
-		return 0;
-
-	/*
-	 * Update power mode time
-	 */
-	update_power_mode_time();
-	power_mode = pm;
-
-	/*
-	 * Change the power mode.
-	 */
-	switch (pm) {
-	case LAVD_PM_PERFORMANCE:
-		no_core_compaction = true;
-		no_freq_scaling = true;
-		no_prefer_turbo_core = false;
-		is_powersave_mode = false;
-
-		/*
-		 * Since the core compaction becomes off, we need to
-		 * reinitialize the active and overflow cpumask for performance
-		 * mode.
-		 *
-		 * Note that a verifier in an old kernel does not allow calling
-		 * bpf_cpumask_set_cpu(), so we defer the actual update to our
-		 * timer handler, update_sys_stat().
-		 */
-		reinit_cpumask_for_performance = true;
-		debugln("Set the scheduler's power profile to performance mode: %d", util);
-		break;
-	case LAVD_PM_BALANCED:
-		no_core_compaction = false;
-		no_freq_scaling = false;
-		no_prefer_turbo_core = false;
-		is_powersave_mode = false;
-		reinit_cpumask_for_performance = false;
-		debugln("Set the scheduler's power profile to balanced mode: %d", util);
-		break;
-	case LAVD_PM_POWERSAVE:
-		no_core_compaction = false;
-		no_freq_scaling = false;
-		no_prefer_turbo_core = true;
-		is_powersave_mode = true;
-		reinit_cpumask_for_performance = false;
-		debugln("Set the scheduler's power profile to power-save mode: %d", util);
-		break;
-	default:
-		return -EINVAL;
-	}
-
-	return 0;
-}
-
-static int do_autopilot(void)
-{
-	struct sys_stat *stat_cur = get_sys_stat_cur();
-
-	/*
-	 * If the CPU utiulization is very low (say <= 5%), it means high
-	 * performance is not required. We run the scheduler in powersave mode
-	 * to save energy consumption.
-	 */
-	if (stat_cur->util <= LAVD_AP_LOW_UTIL)
-		return do_set_power_profile(LAVD_PM_POWERSAVE, stat_cur->util);
-
-	/*
-	 * If the CPU utiulization is moderate (say > 5%, <= 30%), we run the
-	 * scheduler in balanced mode. Actually, balanced mode can save energy
-	 * consumption only under moderate CPU load.
-	 */
-	if (stat_cur->util <= LAVD_AP_HIGH_UTIL)
-		return do_set_power_profile(LAVD_PM_BALANCED, stat_cur->util);
-
-	/*
-	 * If the CPU utilization is high enough (say > 30%), we run the
-	 * scheduler in performance mode. The system indeed needs perrformance
-	 * also there is little energy benefit even under balanced mode anyway.
-	 */
-	return do_set_power_profile(LAVD_PM_PERFORMANCE, stat_cur->util);
-}
-
-static void update_thr_perf_cri(void)
-{
-	struct sys_stat *stat_cur = get_sys_stat_cur();
-	u32 little_core_ratio, delta, diff, thr;
-
-	if (no_core_compaction || !have_little_core)
-		cur_big_core_ratio = default_big_core_ratio;
-
-	/*
-	 * If all active cores are big, all tasks should run on the big cores.
-	 */
-	if (cur_big_core_ratio == 1000) {
-		stat_cur->thr_perf_cri = 0;
-		return;
-	}
-
-	/*
-	 * We approximate the distribution of performance criticality of tasks
-	 * using min, avg, and max performance criticality of a given period.
-	 *
-	 *   min_perf_cri
-	 *   |         avg_perf_cri
-	 *   |         |                       max_perf_cri
-	 *   |         |                       |
-	 *   <--------><----------------------->
-	 *
-	 * The half of compute capacity should be assigned to the below average
-	 * tasks (< avg_perf_cri), and the other half should assigned to the
-	 * above average tasks (>= avg_perf_cri).
-	 *
-	 *   <------------><------------------->
-	 *   |            |                    |
-	 *   |            |                    1000
-	 *   |            1000 - big_core_ratio (i.e., little_core_ratio)
-	 *   0
-	 */
-	little_core_ratio = 1000 - cur_big_core_ratio;
-	if (little_core_ratio < 500) {
-		/*
-		 *   min_perf_cri
-		 *   |         avg_perf_cri
-		 *   |         |                       max_perf_cri
-		 *   |         |                       |
-		 *   <--------><----------------------->
-		 *
-		 *   <-///-><-------------------------->
-		 *   |     |                           |
-		 *   |     |                           1000
-		 *   |     little_core_ratio
-		 *   0
-		 */
-		delta = stat_cur->avg_perf_cri - stat_cur->min_perf_cri;
-		diff = (delta * little_core_ratio) / 1000;
-		thr = diff + stat_cur->min_perf_cri;
-	}
-	else {
-		/*
-		 *   min_perf_cri
-		 *   |         avg_perf_cri
-		 *   |         |                       max_perf_cri
-		 *   |         |                       |
-		 *   <--------><----------------------->
-		 *
-		 *   <---------------------><-////////->
-		 *   |                     |           |
-		 *   |                     |           1000
-		 *   |                     little_core_ratio
-		 *   0
-		 */
-		delta = stat_cur->max_perf_cri - stat_cur->avg_perf_cri;
-		diff = (delta * cur_big_core_ratio) / 1000;
-		thr = stat_cur->max_perf_cri - diff;
-	}
-
-	stat_cur->thr_perf_cri = thr;
-}
-
-static void update_sys_stat(void)
-{
-	do_update_sys_stat();
-
-	if (is_autopilot_on)
-		do_autopilot();
-
-	if (!no_core_compaction)
-		do_core_compaction();
-
-	update_thr_perf_cri();
-
-	if (reinit_cpumask_for_performance) {
-		reinit_cpumask_for_performance = false;
-		reinit_active_cpumask_for_performance();
-	}
-}
-
-static int update_timer_cb(void *map, int *key, struct bpf_timer *timer)
-{
-	int err;
-
-	update_sys_stat();
-
-	err = bpf_timer_start(timer, LAVD_SYS_STAT_INTERVAL_NS, 0);
-	if (err)
-		scx_bpf_error("Failed to arm update timer");
-
-	return 0;
-}
 
 static u32 calc_greedy_ratio(struct task_ctx *taskc)
 {
@@ -1275,12 +233,6 @@ static u32 calc_greedy_ratio(struct task_ctx *taskc)
 out:
 	taskc->greedy_ratio = max(ratio, 1);
 	return ratio;
-}
-
-static u16 get_nice_prio(struct task_struct *p)
-{
-	u16 prio = p->static_prio - MAX_RT_PRIO; /* [0, 40) */
-	return prio;
 }
 
 static u64 calc_runtime_factor(u64 runtime)
@@ -1996,316 +948,6 @@ static void calc_when_to_run(struct task_struct *p, struct task_ctx *taskc,
 	WRITE_ONCE(taskc->vdeadline_log_clk, vlc);
 }
 
-static u64 get_est_stopping_time(struct task_ctx *taskc)
-{
-	return bpf_ktime_get_ns() + taskc->run_time_ns;
-}
-
-static int comp_preemption_info(struct preemption_info *prm_a,
-				struct preemption_info *prm_b)
-{
-	/*
-	 * Check if one's latency priority _or_ deadline is smaller or not.
-	 */
-	if ((prm_a->lat_cri < prm_b->lat_cri) ||
-	    (prm_a->stopping_tm_est_ns < prm_b->stopping_tm_est_ns))
-		return -1;
-	if ((prm_a->lat_cri > prm_b->lat_cri) ||
-	    (prm_a->stopping_tm_est_ns > prm_b->stopping_tm_est_ns))
-		return 1;
-	return 0;
-}
-
-static  bool can_task1_kick_task2(struct preemption_info *prm_task1,
-				  struct preemption_info *prm_task2)
-{
-	return comp_preemption_info(prm_task1, prm_task2) < 0;
-}
-
-static  bool can_cpu1_kick_cpu2(struct preemption_info *prm_cpu1,
-				struct preemption_info *prm_cpu2,
-				struct cpu_ctx *cpuc2)
-{
-	/*
-	 * Set a CPU information
-	 */
-	prm_cpu2->stopping_tm_est_ns = cpuc2->stopping_tm_est_ns;
-	prm_cpu2->lat_cri = cpuc2->lat_cri;
-	prm_cpu2->cpuc = cpuc2;
-	prm_cpu2->last_kick_clk = cpuc2->last_kick_clk;
-
-	/*
-	 * If that CPU runs a lower priority task, that's a victim
-	 * candidate.
-	 */
-	return can_task1_kick_task2(prm_cpu1, prm_cpu2);
-}
-
-static bool is_worth_kick_other_task(struct task_ctx *taskc)
-{
-	/*
-	 * The scx_bpf_kick_cpu() used for preemption is expensive as an IPI is
-	 * involved. Hence, we first judiciously check whether it is worth
-	 * trying to victimize another CPU as the current task is urgent
-	 * enough.
-	 */
-	struct sys_stat *stat_cur = get_sys_stat_cur();
-
-	return (taskc->lat_cri >= stat_cur->thr_lat_cri);
-}
-
-static bool can_cpu_be_kicked(u64 now, struct cpu_ctx *cpuc)
-{
-	return cpuc->is_online &&
-	       (now - cpuc->last_kick_clk) >= LAVD_PREEMPT_KICK_MARGIN;
-}
-
-static struct cpu_ctx *find_victim_cpu(const struct cpumask *cpumask,
-				       struct task_ctx *taskc,
-				       u64 *p_old_last_kick_clk)
-{
-	/*
-	 * We see preemption as a load-balancing problem. In a system with N
-	 * CPUs, ideally, the top N tasks with the highest latency priorities
-	 * should run on the N CPUs all the time. This is the same as the
-	 * load-balancing problem; the load-balancing problem finds a least
-	 * loaded server, and the preemption problem finds a CPU running a
-	 * least latency critical task. Hence, we use the 'power of two random
-	 * choices' technique.
-	 */
-	u64 now = bpf_ktime_get_ns();
-	struct cpu_ctx *cpuc;
-	struct preemption_info prm_task, prm_cpus[2], *victim_cpu;
-	int cpu, nr_cpus;
-	int i, v = 0, cur_cpu = bpf_get_smp_processor_id();
-	int ret;
-
-	/*
-	 * Get task's preemption information for comparison.
-	 */
-	prm_task.stopping_tm_est_ns = get_est_stopping_time(taskc) +
-				      LAVD_PREEMPT_KICK_MARGIN;
-	prm_task.lat_cri = taskc->lat_cri;
-	prm_task.cpuc = cpuc = get_cpu_ctx();
-	if (!cpuc) {
-		scx_bpf_error("Failed to lookup the current cpu_ctx");
-		goto null_out;
-	}
-	prm_task.last_kick_clk = cpuc->last_kick_clk;
-
-	/*
-	 * First, test the current CPU since it can skip the expensive IPI.
-	 */
-	if (can_cpu_be_kicked(now, cpuc) &&
-	    bpf_cpumask_test_cpu(cur_cpu, cpumask) &&
-	    can_cpu1_kick_cpu2(&prm_task, &prm_cpus[0], cpuc)) {
-		victim_cpu = &prm_task;
-		goto bingo_out;
-	}
-
-	/*
-	 * If the current CPU cannot be a victim, let's check if it is worth to
-	 * try to kick other CPU at the expense of IPI.
-	 */
-	if (!is_worth_kick_other_task(taskc))
-		goto null_out;
-
-	/*
-	 * Randomly find _two_ CPUs that run lower-priority tasks than @p. To
-	 * traverse CPUs in a random order, we start from a random CPU ID in a
-	 * random direction (left or right). The random-order traversal helps
-	 * to mitigate the thundering herd problem. Otherwise, all CPUs may end
-	 * up finding the same victim CPU.
-	 *
-	 * In the worst case, the current logic traverses _all_ CPUs. It would
-	 * be too expensive to perform every task queue. We need to revisit
-	 * this if the traversal cost becomes problematic.
-	 */
-	barrier();
-	nr_cpus = bpf_cpumask_weight(cpumask);
-	bpf_for(i, 0, nr_cpus) {
-		/*
-		 * Decide a CPU ID to examine.
-		 */
-		cpu = bpf_cpumask_any_distribute(cpumask);
-
-		if (cpu >= nr_cpu_ids || cur_cpu == cpu)
-			continue;
-
-		/*
-		 * Check whether that CPU is qualified to run @p.
-		 */
-		cpuc = get_cpu_ctx_id(cpu);
-		if (!cpuc) {
-			scx_bpf_error("Failed to lookup cpu_ctx: %d", cpu);
-			goto null_out;
-		}
-
-		if (!can_cpu_be_kicked(now, cpuc))
-			continue;
-
-		/*
-		 * If that CPU runs a lower priority task, that's a victim
-		 * candidate.
-		 */
-		ret = can_cpu1_kick_cpu2(&prm_task, &prm_cpus[v], cpuc);
-		if (ret == true && ++v >= 2)
-			break;
-	}
-
-	/*
-	 * Choose a final victim CPU.
-	 */
-	switch(v) {
-	case 2:	/* two dandidates */
-		victim_cpu = can_task1_kick_task2(&prm_cpus[0], &prm_cpus[1]) ?
-				&prm_cpus[0] : &prm_cpus[1];
-		goto bingo_out;
-	case 1:	/* one candidate */
-		victim_cpu = &prm_cpus[0];
-		goto bingo_out;
-	case 0:	/* no candidate */
-		goto null_out;
-	default:/* something wrong */
-		goto null_out;
-	}
-
-bingo_out:
-	taskc->victim_cpu = victim_cpu->cpuc->cpu_id;
-	*p_old_last_kick_clk = victim_cpu->last_kick_clk;
-	return victim_cpu->cpuc;
-
-null_out:
-	taskc->victim_cpu = (s32)LAVD_CPU_ID_NONE;
-	return NULL;
-}
-
-static bool kick_cpu(struct cpu_ctx *victim_cpuc, u64 victim_last_kick_clk)
-{
-	/*
-	 * If the current CPU is a victim, we just reset the current task's
-	 * time slice as an optimization. Othewise, kick the remote CPU for
-	 * preemption.
-	 *
-	 * Kicking the victim CPU does _not_ guarantee that task @p will run on
-	 * that CPU. Enqueuing @p to the global queue is one operation, and
-	 * kicking the victim is another asynchronous operation. However, it is
-	 * okay because, anyway, the victim CPU will run a higher-priority task
-	 * than @p.
-	 */
-	if (bpf_get_smp_processor_id() == victim_cpuc->cpu_id) {
-		struct task_struct *tsk = bpf_get_current_task_btf();
-		tsk->scx.slice = 0;
-		return true;
-	}
-
-	/*
-	 * Kick the remote victim CPU if it is not victimized yet by another
-	 * concurrent kick task.
-	 */
-	bool ret = __sync_bool_compare_and_swap(&victim_cpuc->last_kick_clk,
-						victim_last_kick_clk,
-						bpf_ktime_get_ns());
-	if (ret)
-		scx_bpf_kick_cpu(victim_cpuc->cpu_id, SCX_KICK_PREEMPT);
-
-	return ret;
-}
-
-static bool try_find_and_kick_victim_cpu(struct task_struct *p,
-					 struct task_ctx *taskc,
-					 struct cpu_ctx *cpuc_cur,
-					 u64 dsq_id)
-{
-	struct bpf_cpumask *cd_cpumask, *cpumask;
-	struct cpdom_ctx *cpdomc;
-	struct cpu_ctx *victim_cpuc;
-	u64 victim_last_kick_clk;
-	bool ret = false;
-
-	/*
-	 * Prepare a cpumak so we find a victim @p's compute domain.
-	 */
-	cpumask = cpuc_cur->tmp_t_mask;
-	cpdomc = MEMBER_VPTR(cpdom_ctxs, [dsq_id]);
-	cd_cpumask = MEMBER_VPTR(cpdom_cpumask, [dsq_id]);
-	if (!cpdomc || !cd_cpumask || !cpumask)
-		return false;
-
-	bpf_cpumask_and(cpumask, cast_mask(cd_cpumask), p->cpus_ptr);
-
-	/*
-	 * Find a victim CPU among CPUs that run lower-priority tasks.
-	 */
-	victim_cpuc = find_victim_cpu(cast_mask(cpumask), taskc, &victim_last_kick_clk);
-
-	/*
-	 * If a victim CPU is chosen, preempt the victim by kicking it.
-	 */
-	if (victim_cpuc)
-		ret = kick_cpu(victim_cpuc, victim_last_kick_clk);
-
-	if (!ret)
-		taskc->victim_cpu = (s32)LAVD_CPU_ID_NONE;
-
-	return ret;
-}
-
-static bool try_yield_current_cpu(struct task_struct *p_run,
-				  struct cpu_ctx *cpuc_run,
-				  struct task_ctx *taskc_run)
-{
-	struct task_struct *p_wait;
-	struct task_ctx *taskc_wait;
-	struct preemption_info prm_run, prm_wait;
-	s32 cpu_id = scx_bpf_task_cpu(p_run), wait_vtm_cpu_id;
-	bool ret = false;
-
-	/*
-	 * If there is a higher priority task waiting on the global rq, the
-	 * current running task yield the CPU by shrinking its time slice to
-	 * zero.
-	 */
-	prm_run.stopping_tm_est_ns = taskc_run->last_running_clk +
-				     taskc_run->run_time_ns -
-				     LAVD_PREEMPT_TICK_MARGIN;
-	prm_run.lat_cri = taskc_run->lat_cri;
-
-	bpf_rcu_read_lock();
-	bpf_for_each(scx_dsq, p_wait, cpuc_run->cpdom_id, 0) {
-		taskc_wait = get_task_ctx(p_wait);
-		if (!taskc_wait)
-			break;
-
-		wait_vtm_cpu_id = taskc_wait->victim_cpu;
-		if (wait_vtm_cpu_id != (s32)LAVD_CPU_ID_NONE)
-			break;
-
-		prm_wait.stopping_tm_est_ns = get_est_stopping_time(taskc_wait);
-		prm_wait.lat_cri = taskc_wait->lat_cri;
-
-		if (can_task1_kick_task2(&prm_wait, &prm_run)) {
-			/*
-			 * The atomic CAS guarantees only one task yield its
-			 * CPU for the waiting task.
-			 */
-			ret = __sync_bool_compare_and_swap(
-					&taskc_wait->victim_cpu,
-					(s32)LAVD_CPU_ID_NONE, cpu_id);
-			if (ret)
-				p_run->scx.slice = 0;
-		}
-
-		/*
-		 * Test only the first entry on the DSQ.
-		 */
-		break;
-	}
-	bpf_rcu_read_unlock();
-
-	return ret;
-}
-
 static u64 find_proper_dsq(struct task_ctx *taskc, struct cpu_ctx *cpuc)
 {
 	struct sys_stat *stat_cur = get_sys_stat_cur();
@@ -2390,25 +1032,6 @@ void BPF_STRUCT_OPS(lavd_enqueue, struct task_struct *p, u64 enq_flags)
 	 */
 	scx_bpf_dispatch_vtime(p, dsq_id, p->scx.slice,
 			       taskc->vdeadline_log_clk, enq_flags);
-}
-
-static bool use_full_cpus(void)
-{
-	struct sys_stat *stat_cur = get_sys_stat_cur();
-	return (stat_cur->nr_active + LAVD_CC_NR_OVRFLW) >= nr_cpus_onln;
-}
-
-static u64 pick_any_bit(u64 bitmap, u64 nuance)
-{
-	u64 i, pos;
-
-	bpf_for(i, 0, 64) {
-		pos = (i + nuance) % 64;
-		if (bitmap & (1LLU << pos))
-			return pos;
-	}
-
-	return -ENOENT;
 }
 
 static bool consume_dsq(s32 cpu, u64 dsq_id, u64 now)
@@ -2661,95 +1284,6 @@ consume_out:
 		}
 		prev->scx.slice = calc_time_slice(prev, taskc);
 	}
-}
-
-static int calc_cpuperf_target(struct sys_stat *stat_cur,
-			       struct task_ctx *taskc, struct cpu_ctx *cpuc)
-{
-	u64 max_load, cpu_load;
-	u32 cpuperf_target;
-
-	if (!stat_cur || !taskc || !cpuc)
-		return -EINVAL;
-
-	if (no_freq_scaling) {
-		cpuc->cpuperf_task = SCX_CPUPERF_ONE;
-		cpuc->cpuperf_avg = SCX_CPUPERF_ONE;
-		return 0;
-	}
-
-	/*
-	 * We determine the clock frequency of a CPU using two factors: 1) the
-	 * current CPU utilization (cpuc->util) and 2) the current task's
-	 * performance criticality (taskc->perf_cri) compared to the
-	 * system-wide average performance criticality
-	 * (stat_cur->thr_perf_cri).
-	 *
-	 * When a current CPU utilization is 85% and the current task's
-	 * performance criticality is the same as the system-wide average
-	 * criticality, we set the target CPU frequency to the maximum.
-	 *
-	 * In other words, even if CPU utilization is not so high, the target
-	 * CPU frequency could be high when the task's performance criticality
-	 * is high enough (i.e., boosting CPU frequency). On the other hand,
-	 * the target CPU frequency could be low even if CPU utilization is
-	 * high when a non-performance-critical task is running (i.e.,
-	 * deboosting CPU frequency).
-	 */
-	max_load = stat_cur->thr_perf_cri * LAVD_CPU_UTIL_MAX_FOR_CPUPERF;
-	cpu_load = taskc->perf_cri * cpuc->util;
-	cpuperf_target = (cpu_load * SCX_CPUPERF_ONE) / max_load;
-	cpuperf_target = min(cpuperf_target, SCX_CPUPERF_ONE);
-
-	cpuc->cpuperf_task = cpuperf_target;
-	cpuc->cpuperf_avg = calc_avg32(cpuc->cpuperf_avg, cpuperf_target);
-	return 0;
-}
-
-static bool try_increase_cpuperf_target(struct cpu_ctx *cpuc)
-{
-	/*
-	 * When a task becomes running, update CPU's performance target only
-	 * when the current task's target performance is higher. This helps
-	 * rapidly adopt workload changes by rapidly increasing CPU's
-	 * performance target.
-	 */
-	u32 target;
-
-	if (!cpuc)
-		return false;
-
-	target = max(cpuc->cpuperf_task, cpuc->cpuperf_avg);
-	if (cpuc->cpuperf_cur < target) {
-		cpuc->cpuperf_cur = target;
-		scx_bpf_cpuperf_set(cpuc->cpu_id, target);
-		return true;
-	}
-
-	return false;
-}
-
-static bool try_decrease_cpuperf_target(struct cpu_ctx *cpuc)
-{
-	/*
-	 * Upon every tick interval, we try to decrease the CPU's performance
-	 * target if the current one is higher than both the current task's
-	 * target and EWMA of past targets. This helps gradually adopt workload
-	 * changes upon sudden down falls.
-	 */
-	u32 target;
-
-	if (!cpuc)
-		return false;
-
-	target = max(cpuc->cpuperf_task, cpuc->cpuperf_avg);
-	if (cpuc->cpuperf_cur != target) {
-		cpuc->cpuperf_cur = target;
-		scx_bpf_cpuperf_set(cpuc->cpu_id, target);
-		return true;
-	}
-
-	return false;
 }
 
 void BPF_STRUCT_OPS(lavd_tick, struct task_struct *p_run)
@@ -3074,36 +1608,6 @@ void BPF_STRUCT_OPS(lavd_update_idle, s32 cpu, bool idle)
 	}
 }
 
-static void set_on_core_type(struct task_ctx *taskc,
-				 const struct cpumask *cpumask)
-{
-	bool on_big = false, on_little = false;
-	struct cpu_ctx *cpuc;
-	int cpu;
-
-	bpf_for(cpu, 0, nr_cpu_ids) {
-		if (!bpf_cpumask_test_cpu(cpu, cpumask))
-			continue;
-
-		cpuc = get_cpu_ctx_id(cpu);
-		if (!cpuc) {
-			scx_bpf_error("Failed to look up cpu_ctx: %d", cpu);
-			return;
-		}
-
-		if (cpuc->big_core)
-			on_big = true;
-		else
-			on_little = true;
-
-		if (on_big && on_little)
-			break;
-	}
-
-	WRITE_ONCE(taskc->on_big, on_big);
-	WRITE_ONCE(taskc->on_little, on_little);
-}
-
 void BPF_STRUCT_OPS(lavd_set_cpumask, struct task_struct *p,
 		    const struct cpumask *cpumask)
 {
@@ -3116,21 +1620,6 @@ void BPF_STRUCT_OPS(lavd_set_cpumask, struct task_struct *p,
 	}
 
 	set_on_core_type(taskc, cpumask);
-}
-
-static void init_task_ctx(struct task_struct *p, struct task_ctx *taskc)
-{
-	struct sys_stat *stat_cur = get_sys_stat_cur();
-	u64 now = bpf_ktime_get_ns();
-
-	memset(taskc, 0, sizeof(*taskc));
-	taskc->last_running_clk = now; /* for run_time_ns */
-	taskc->last_stopping_clk = now; /* for run_time_ns */
-	taskc->run_time_ns = LAVD_SLICE_MAX_NS;
-	taskc->victim_cpu = (s32)LAVD_CPU_ID_NONE;
-	taskc->svc_time = stat_cur->avg_svc_time * LAVD_NEW_PROC_PENALITY;
-
-	set_on_core_type(taskc, p->cpus_ptr);
 }
 
 void BPF_STRUCT_OPS(lavd_enable, struct task_struct *p)
@@ -3147,6 +1636,21 @@ void BPF_STRUCT_OPS(lavd_enable, struct task_struct *p)
 	}
 
 	taskc->svc_time = READ_ONCE(cur_svc_time);
+}
+
+static void init_task_ctx(struct task_struct *p, struct task_ctx *taskc)
+{
+	struct sys_stat *stat_cur = get_sys_stat_cur();
+	u64 now = bpf_ktime_get_ns();
+
+	memset(taskc, 0, sizeof(*taskc));
+	taskc->last_running_clk = now; /* for run_time_ns */
+	taskc->last_stopping_clk = now; /* for run_time_ns */
+	taskc->run_time_ns = LAVD_SLICE_MAX_NS;
+	taskc->victim_cpu = (s32)LAVD_CPU_ID_NONE;
+	taskc->svc_time = stat_cur->avg_svc_time * LAVD_NEW_PROC_PENALITY;
+
+	set_on_core_type(taskc, p->cpus_ptr);
 }
 
 s32 BPF_STRUCT_OPS(lavd_init_task, struct task_struct *p,
@@ -3257,83 +1761,6 @@ static int init_cpumasks(void)
 	scx_bpf_put_cpumask(online_cpumask);
 
 out:
-	bpf_rcu_read_unlock();
-	return err;
-}
-
-static u16 get_cpuperf_cap(s32 cpu)
-{
-	if (cpu >= 0 && cpu < nr_cpu_ids && cpu < LAVD_CPU_ID_MAX)
-		return __cpu_capacity_hint[cpu];
-
-	debugln("Infeasible CPU id: %d", cpu);
-	return 0;
-}
-
-static u16 get_cputurbo_cap(void)
-{
-	u16 turbo_cap = 0;
-	int nr_turbo = 0, cpu;
-
-	/*
-	 * Find the maximum CPU frequency
-	 */
-	for (cpu = 0; cpu < nr_cpu_ids && cpu < LAVD_CPU_ID_MAX; cpu++) {
-		if (__cpu_capacity_hint[cpu] > turbo_cap) {
-			turbo_cap = __cpu_capacity_hint[cpu];
-			nr_turbo++;
-		}
-	}
-
-	/*
-	 * If all CPU's frequencies are the same, ignore the turbo.
-	 */
-	if (nr_turbo <= 1)
-		turbo_cap = 0;
-
-	return turbo_cap;
-}
-
-static int reinit_active_cpumask_for_performance(void)
-{
-	struct cpu_ctx *cpuc;
-	struct bpf_cpumask *active, *ovrflw;
-	int cpu, err = 0;
-
-	barrier();
-	bpf_rcu_read_lock();
-
-	/*
-	 * Prepare cpumasks.
-	 */
-	active  = active_cpumask;
-	ovrflw  = ovrflw_cpumask;
-	if (!active || !ovrflw) {
-		scx_bpf_error("Failed to prepare cpumasks.");
-		err = -ENOMEM;
-		goto unlock_out;
-	}
-
-
-	/*
-	 * Once core compaction becomes off in performance mode,
-	 * reinitialize active/overflow cpumasks to reflect the mode change.
-	 */
-	bpf_for(cpu, 0, nr_cpu_ids) {
-		cpuc = get_cpu_ctx_id(cpu);
-		if (!cpuc) {
-			scx_bpf_error("Failed to lookup cpu_ctx: %d", cpu);
-			err = -ESRCH;
-			goto unlock_out;
-		}
-
-		if (cpuc->big_core)
-			bpf_cpumask_set_cpu(cpu, active);
-		else
-			bpf_cpumask_set_cpu(cpu, ovrflw);
-	}
-
-unlock_out:
 	bpf_rcu_read_unlock();
 	return err;
 }
@@ -3499,52 +1926,6 @@ unlock_out:
 	return err;
 }
 
-static s32 init_sys_stat(u64 now)
-{
-	struct bpf_timer *timer;
-	u32 key = 0;
-	int err;
-
-	memset(__sys_stats, 0, sizeof(__sys_stats));
-	__sys_stats[0].last_update_clk = now;
-	__sys_stats[1].last_update_clk = now;
-	__sys_stats[0].nr_active = nr_cpus_big;
-	__sys_stats[1].nr_active = nr_cpus_big;
-
-	timer = bpf_map_lookup_elem(&update_timer, &key);
-	if (!timer) {
-		scx_bpf_error("Failed to lookup update timer");
-		return -ESRCH;
-	}
-	bpf_timer_init(timer, &update_timer, CLOCK_BOOTTIME);
-	bpf_timer_set_callback(timer, update_timer_cb);
-	err = bpf_timer_start(timer, LAVD_SYS_STAT_INTERVAL_NS, 0);
-	if (err) {
-		scx_bpf_error("Failed to arm update timer");
-		return err;
-	}
-
-	return 0;
-}
-
-static void init_autopilot_low_util(void)
-{
-	if (nr_cpus_big < nr_cpus_onln) {
-		/*
-		 * When there are little cores, we move up to the balanced mode
-		 * if one little core is fully utilized.
-		 */
-		LAVD_AP_LOW_UTIL = 1000 / nr_cpus_onln;
-	}
-	else {
-		/*
-		 * When there are only big cores, we move up to the balanced
-		 * mode if two big cores are fully utilized.
-		 */
-		LAVD_AP_LOW_UTIL = (2 * 1000) / nr_cpus_onln;
-	}
-}
-
 s32 BPF_STRUCT_OPS_SLEEPABLE(lavd_init)
 {
 	u64 now = bpf_ktime_get_ns();
@@ -3599,12 +1980,6 @@ s32 BPF_STRUCT_OPS_SLEEPABLE(lavd_init)
 void BPF_STRUCT_OPS(lavd_exit, struct scx_exit_info *ei)
 {
 	UEI_RECORD(uei, ei);
-}
-
-SEC("syscall")
-int set_power_profile(struct power_arg *input)
-{
-	return do_set_power_profile(input->power_mode, 0);
 }
 
 SCX_OPS_DEFINE(lavd_ops,
